requirements.txt
playwright
psycopg2-binary
beautifulsoup4